admin can not leave group, more controller tests

+++ app/controllers/memberships_controller.rb
@@ -18,7 +18,7 @@ class MembershipsController < ApplicationController
     # Membership.find(params[:id]).destroy
     @board = Board.find(params[:membership][:board_id])
     membership = Membership.find_by(user_id: @current_user.id, board_id: @board.id)
-    membership.destroy
+    membership.destroy if !membership.admin?
     respond_to do |format|
       format.html { redirect_to @board }
       format.js
